changed test case for bts1700

+++ src/plugins/e-acsl/tests/e-acsl-runtime/oracle/gen_bts1700.c
@@ -22,7 +22,7 @@ extern  __attribute__((__FC_BUILTIN__)) void e_acsl_assert(int predicate,
 model __mpz_struct { ℤ n };
 */
 int __fc_random_counter __attribute__((__unused__));
-unsigned long const __fc_rand_max = 32767UL;
+unsigned long const __fc_rand_max = (unsigned long)32767;
 /*@ ghost extern int __fc_heap_status; */
 
 /*@
